requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
httpx
tzdata
